move ensureConnect to shared file
